signal to ValBox is float instead of int

+++ libpyEM/qtgui/valslider.py
@@ -301,7 +301,7 @@ class ValBox(QtGui.QWidget):
 	so for virtually all purposes it could be used as a drop-in replacement.
 	"""
 	enableChanged = QtCore.pyqtSignal(int)
-	valueChanged = QtCore.pyqtSignal(int)
+	valueChanged = QtCore.pyqtSignal(float)
 	textChanged = QtCore.pyqtSignal()
 
 	def __init__(self, parent=None, rng=None, label=None, value=0,labelwidth=30,showenable=-1):
